// frontend/src/backend.d.ts
import type { Principal } from "@icp-sdk/core/principal";
export interface Some<T> {
    __kind__: "Some";
    value: T;
}
export interface None {
    __kind__: "None";
}
export type Option<T> = Some<T> | None;
export interface TransformationInput {
    context: Uint8Array;
    response: http_request_result;
}
export interface CombinedBrowserData {
    apkmirror: Array<BrowserData>;
    browserUpdate: Array<BrowserData>;
    statcounter: Array<BrowserData>;
}
export interface BrowserData {
    source: string;
    name: string;
    version: string;
}
export interface TransformationOutput {
    status: bigint;
    body: Uint8Array;
    headers: Array<http_header>;
}
export interface http_header {
    value: string;
    name: string;
}
export interface http_request_result {
    status: bigint;
    body: Uint8Array;
    headers: Array<http_header>;
}
export interface backendInterface {
    fetchApkMirrorData(): Promise<string>;
    fetchBrowserUpdateData(): Promise<string>;
    fetchStatCounterData(): Promise<string>;
    getAllBrowserData(): Promise<CombinedBrowserData>;
    transform(input: TransformationInput): Promise<TransformationOutput>;
}